Crate massa_ledger

General description

This crate implements a ledger matching addresses to balances, executable bytecode and data. It also provides tools to manipulate ledger entries.

FinalLedger is used as part of FinalState that represents the latest final state of the node (see massa-final-state crate for more details).

FinalLedger representing a ledger at a given slot that was executed as final (see the massa-execution-worker crate for details on execution). Only the execution worker writes into the final ledger.

A note on parallel vs sequential balance

The distinctions between the parallel and the sequential balance of a ledger entry are the following:

the parallel balance can be credited or spent in any slot

the sequential balance can be credited in any slot but only spent in slots form the address’ thread

block produers are credited fees from the sequential balance, and they can ensure that this balance will be available for their block simply by looking for sequential balance spendings within the block’s thread.

Architecture

ledger.rs

Defines the FinalLedger that matches an address to a LedgerEntry (see ledger_entry.rs), and can be manipulated using LedgerChanges (see ledger_changes.rs). The FinalLedger is bootstrapped using tooling available in bootstrap.rs

ledger_entry.rs

Represents an entry in the ledger for a given address. It contains balances, executable bytecode and an arbitrary datastore.

ledger_changes.rs

Represents a list of changes to ledger entries that can be modified, combined or applied to the final ledger.

bootstrap.rs

Provides serializable strucutres and tools for bootstrapping the final ledger.

Test exports

When the crate feature testing is enabled, tooling useful for testing purposes is exported. See test_exports/mod.rs for details.

Structs

FinalLedger

Represents a final ledger associating addresses to their balances, bytecode and data. The final ledger is part of the final state which is attached to a final slot, can be boostrapped and allows others to bootstrap. The ledger size can be very high: it can exceed 1TB. To allow for storage on disk, the ledger uses trees and has O(log(N)) access, insertion and deletion complexity.

FinalLedgerBootstrapState

Represents a snapshot of the ledger state, which is enough to fully bootstrap a FinalLedger

LedgerChanges

represents a list of changes to multiple ledger entries

LedgerConfig

Ledger configuration

LedgerEntry

Structure defining an entry associated to an address in the FinalLedger

Enums

LedgerError

SetOrDelete

Enum representing a set/delete change on a value T

SetOrKeep

represents a set/keep change

SetUpdateOrDelete

Enum representing set/update/delete change on a value T

Traits

Applicable

Trait marking a structure that supports another one (V) being applied to it