Struct massa_ledger::LedgerChanges

pub struct LedgerChanges(pub Map<Address, SetUpdateOrDelete<LedgerEntry, LedgerEntryUpdate>>);

represents a list of changes to multiple ledger entries

Tuple Fields
0: Map<Address, SetUpdateOrDelete<LedgerEntry, LedgerEntryUpdate>>
Implementations

impl LedgerChanges

pub fn get(
&self,
addr: &Address
) -> Option<&SetUpdateOrDelete<LedgerEntry, LedgerEntryUpdate>>

get an item from the LedgerChanges

pub fn get_parallel_balance_or_else<F: FnOnce() -> Option<Amount>>(
&self,
addr: &Address,
f: F
) -> Option<Amount>

Tries to return the parallel balance of an entry or gets it from a function if the entry’s status is unknown.

This function is used as an optimization: if the value can be deduced unambiguously from the LedgerChanges, no need to dig further (for example in the FinalLedger).

Arguments

addr: address for which to get the value

f: fallback function with no arguments and returning Option

Returns

Some(v) if a value is present, where v is a copy of the value

None if the value is absent

f() if the value is unknown

pub fn get_bytecode_or_else<F: FnOnce() -> Option<Vec<u8>>>(
&self,
addr: &Address,
f: F
) -> Option<Vec<u8>>

Tries to return the executable bytecode of an entry or gets it from a function if the entry’s status is unknown.

This function is used as an optimization: if the value can be deduced unambiguously from the LedgerChanges, no need to dig further (for example in the FinalLedger).

Arguments

addr: address for which to get the value

f: fallback function with no arguments and returning Option<Vec>

Returns

Some(v) if a value is present, where v is a copy of the value

None if the value is absent

f() if the value is unknown

pub fn entry_exists_or_else<F: FnOnce() -> bool>(
&self,
addr: &Address,
f: F
) -> bool

Tries to return whether an entry exists or gets the information from a function if the entry’s status is unknown.

This function is used as an optimization: if the result can be deduced unambiguously from the LedgerChanges, no need to dig further (for example in the FinalLedger).

Arguments

addr: address to search for

f: fallback function with no arguments and returning bool

Returns

true if the entry exists

false if the value is absent

f() if the value’s existence is unknown

pub fn set_parallel_balance(&mut self, addr: Address, balance: Amount)

Set the parallel balance of an address. If the address doesn’t exist, its ledger entry is created.

Arguments

addr: target address

balance: parallel balance to set for the provided address

pub fn set_bytecode(&mut self, addr: Address, bytecode: Vec<u8>)

Set the executable bytecode of an address. If the address doesn’t exist, its ledger entry is created.

Parameters

addr: target address

bytecode: executable bytecode to assign to that address

pub fn get_data_entry_or_else<F: FnOnce() -> Option<Vec<u8>>>(
&self,
addr: &Address,
key: &Hash,
f: F
) -> Option<Vec<u8>>

Tries to return a datastore entry for a given address, or gets it from a function if the value’s status is unknown.

This function is used as an optimization: if the result can be deduced unambiguously from the LedgerChanges, no need to dig further (for example in the FinalLedger).

Arguments

addr: target address

key: datastore key

f: fallback function with no arguments and returning Option<Vec>

Returns

Some(v) if the value was found, where v is a copy of the value

None if the value is absent

f() if the value is unknown

pub fn has_data_entry_or_else<F: FnOnce() -> bool>(
&self,
addr: &Address,
key: &Hash,
f: F
) -> bool

Tries to return wherther a datastore entry exists for a given address, or gets it from a function if the datastore entry’s status is unknown.

This function is used as an optimization: if the result can be deduced unambiguously from the LedgerChanges, no need to dig further (for example in the FinalLedger).

Arguments

addr: target address

key: datastore key

f: fallback function with no arguments and returning bool

Returns

true if the ledger entry exists and the key is present in its datastore

false if the ledger entry is absent, or if the key is not in its datastore

f() if the existence of the ledger entry or datastore entry is unknown

pub fn set_data_entry(&mut self, addr: Address, key: Hash, data: Vec<u8>)

Set a datastore entry for a given address.
If the address doesn’t exist, its ledger entry is created. If the datasotre entry exists, its value is replaced, otherwise it is created.

Arguments

addr: target address

key: datastore key

data: datastore value to set