Struct massa_ledger::FinalLedger

pub struct FinalLedger { /* private fields */ }

Represents a final ledger associating addresses to their balances, bytecode and data. The final ledger is part of the final state which is attached to a final slot, can be boostrapped and allows others to bootstrap. The ledger size can be very high: it can exceed 1TB. To allow for storage on disk, the ledger uses trees and has O(log(N)) access, insertion and deletion complexity.

Note: currently the ledger is stored in RAM. TODO put it on the hard drive with cache.

Implementations

impl FinalLedger

pub fn new(config: LedgerConfig) -> Result<Self, LedgerError>

Initializes a new FinalLedger by reading its initial state from file.

pub fn from_bootstrap_state(
config: LedgerConfig,
state: FinalLedgerBootstrapState
) -> Self

Intiialize a FinalLedger from a bootstrap state

TODO: This loads the whole ledger in RAM. Switch to streaming in the future

Arguments

config: ledger config

state: bootstrap state

pub fn get_bootstrap_state(&self) -> FinalLedgerBootstrapState

Gets a snapshot of the ledger to bootstrap other nodes

TODO: This loads the whole ledger in RAM. Switch to streaming in the future

pub fn get_full_entry(&self, addr: &Address) -> Option<LedgerEntry>

Gets a copy of a full ledger entry.

Returns

A clone of the whole LedgerEntry, or None if not found.

TODO: in the future, never manipulate full ledger entries because their datastore can be huge https://github.com/massalabs/massa/issues/2342

pub fn get_parallel_balance(&self, addr: &Address) -> Option<Amount>

Gets the parallel balance of a ledger entry

Returns

The parallel balance, or None if the ledger entry was not found

pub fn get_bytecode(&self, addr: &Address) -> Option<Vec<u8>>

Gets a copy of the bytecode of a ledger entry

Returns

A copy of the found bytecode, or None if the ledger entry was not found

pub fn entry_exists(&self, addr: &Address) -> bool

Checks if a ledger entry exists

Returns

true if it exists, false otherwise.

pub fn get_data_entry(&self, addr: &Address, key: &Hash) -> Option<Vec<u8>>

Gets a copy of the value of a datastore entry for a given address.

Arguments

addr: target address

key: datastore key

Returns

A copy of the datastore value, or None if the ledger entry or datastore entry was not found

pub fn has_data_entry(&self, addr: &Address, key: &Hash) -> bool

Checks for the existence of a datastore entry for a given address.

Arguments

addr: target address

key: datastore key

Returns

true if the datastore entry was found, or false if the ledger entry or datastore entry was not found